start adding helper reefoctr problems of past system of radio (list of lists now, helper needed becuase the civilian and the gaurd will use same code)

+++ Source/K7/K7CombatBase.h
@@ -73,5 +73,5 @@ public:
     FHitResult Fire(FVector Start,FVector Direction);
     // radio
     ARadioItem* radioCur;
-    float getInfoById(uint8 id);
+    float getInfoById(int32 id);
 };
